Remove audio since Tails 6.1 pipewire change
Tails 6.1 has moved to pipewire instead of pulseaudio for sound. Audio is not working in VMs. Disable audio in VM settings to launch VM. I have no time to solve this for currently. VirtualBox might not yet be easily compatible with pipewire. Installing pulseaudio package(s) and using a pulseaudio bridge or other tricks might be a solution.